test(Python): Add CRS tests
